fix: unpack terraform binaries directly to /usr/bin directory
The `clowdhaus/terraform-composite-actions/pre-commit` action currently
downloads and unpacks terraform and related binaries into the current
working directory and then subsequently moves them into /usr/bin. In a
somewhat contrived (but real) scenario where a GitHub repo already has
a `terraform` directory, unzip will fail since we cannot unzip a binary
on top of a directory with the same name.

With this change, binaries are unpacked directly to the /usr/bin
directory via either

* unzip -qq <source.zip> <target file> -d /usr/bin
* tar -xzf <source.tgz> <target file> -C /usr/bin

These are unpacked with `sudo` in order to be able to write to the
/usr/bin directory. I've also removed the `chmod +x` step for
terraform-docs as it appears to already have the execute bit set on it.