fix: untar re-order untar args so terraform-docs is installed
In a previous commit, I changed the way that various terraform binaries
were unpacked so that they would install directly into /usr/bin rather
than staging in cwd and then moved. Unfortunately this did not work for
terraform-docs as I put the arguments in the wrong order. The -C flag is
order-sensitive as described in the man page, and affects all options
that follow.

This fix moves the `-C /usr/bin/` flag ahead of the file being untar'd
so that it is correctly changes directories prior to terraform-docs
being unpacked.